harfbuzz_rs::font_funcs

Trait FontFuncs

pub trait FontFuncs {
    // Provided methods
    fn get_font_h_extents(&self, font: &Font<'_>) -> Option<FontExtents> { ... }
    fn get_font_v_extents(&self, font: &Font<'_>) -> Option<FontExtents> { ... }
    fn get_nominal_glyph(&self, font: &Font<'_>, unicode: char) -> Option<Glyph> { ... }
    fn get_variation_glyph(
        &self,
        font: &Font<'_>,
        unicode: char,
        variation_sel: char,
    ) -> Option<Glyph> { ... }
    fn get_glyph_h_advance(&self, font: &Font<'_>, glyph: Glyph) -> Position { ... }
    fn get_glyph_v_advance(&self, font: &Font<'_>, glyph: Glyph) -> Position { ... }
    fn get_glyph_h_origin(
        &self,
        font: &Font<'_>,
        glyph: Glyph,
    ) -> Option<(Position, Position)> { ... }
    fn get_glyph_v_origin(
        &self,
        font: &Font<'_>,
        glyph: Glyph,
    ) -> Option<(Position, Position)> { ... }
    fn get_glyph_extents(
        &self,
        font: &Font<'_>,
        glyph: Glyph,
    ) -> Option<GlyphExtents> { ... }
    fn get_glyph_contour_point(
        &self,
        font: &Font<'_>,
        glyph: Glyph,
        point_index: u32,
    ) -> Option<(Position, Position)> { ... }
    fn get_glyph_name(&self, font: &Font<'_>, glyph: Glyph) -> Option<String> { ... }
    fn get_glyph_from_name(&self, font: &Font<'_>, name: &str) -> Option<Glyph> { ... }
}

This Trait specifies the font callbacks that harfbuzz uses for its shaping.

Custom font data providers can implement this trait to supply the HarfBuzz font shaping engine with the necessary information it needs during the shaping process.

Users wishing to use a font data provider that implements FontFuncs can tell HarfBuzz to use the data provider by calling Font::set_font_funcs() on a Font.

Usually this is not necessary, since HarfBuzz comes by default with an integrated implementation of FontFuncs that reads the required data from an OpenType font file.

When implementing this trait, you may only want to override some of the functions, while the rest of the functions should fall back to the HarfBuzz provided font-funcs. This is possible through the use of subfonts.

Note that if a Font is created directly from a face, its parent is the empty Font which returns null values for every font func.

Provided Methods

fn get_font_h_extents(&self, font: &Font<'_>) -> Option<FontExtents>

fn get_font_v_extents(&self, font: &Font<'_>) -> Option<FontExtents>

fn get_nominal_glyph(&self, font: &Font<'_>, unicode: char) -> Option<Glyph>

fn get_variation_glyph( &self, font: &Font<'_>, unicode: char, variation_sel: char, ) -> Option<Glyph>

fn get_glyph_h_advance(&self, font: &Font<'_>, glyph: Glyph) -> Position

fn get_glyph_v_advance(&self, font: &Font<'_>, glyph: Glyph) -> Position

fn get_glyph_h_origin( &self, font: &Font<'_>, glyph: Glyph, ) -> Option<(Position, Position)>

fn get_glyph_v_origin( &self, font: &Font<'_>, glyph: Glyph, ) -> Option<(Position, Position)>

fn get_glyph_extents( &self, font: &Font<'_>, glyph: Glyph, ) -> Option<GlyphExtents>

fn get_glyph_contour_point( &self, font: &Font<'_>, glyph: Glyph, point_index: u32, ) -> Option<(Position, Position)>

fn get_glyph_name(&self, font: &Font<'_>, glyph: Glyph) -> Option<String>

fn get_glyph_from_name(&self, font: &Font<'_>, name: &str) -> Option<Glyph>