harfbuzz_rs

Struct Face

pub struct Face<'a> { /* private fields */ }

A wrapper around hb_face_t.

An excerpt from harfbuzz documentation:

Font face is objects represent a single face in a font family. More exactly, a font face represents a single face in a binary font file. Font faces are typically built from a binary blob and a face index. Font faces are used to create fonts.

Implementations

impl<'a> Face<'a>

pub fn new<T: Into<Shared<Blob<'a>>>>(data: T, index: u32) -> Owned<Face<'a>>

Create a new Face from the data.

If data is not a valid font then this function returns the empty face.

pub fn empty() -> Owned<Face<'static>>

Returns a “null” face.

pub fn from_file<P: AsRef<Path>>( path: P, index: u32, ) -> Result<Owned<Face<'static>>>

Create a new face from the contents of the file at path.

This function reads the contents of the file at path into memory, creates a Blob and then calls Face::new.

See also the discussion in Blob::from_file.

pub fn from_bytes<'b>(bytes: &'b [u8], index: u32) -> Owned<Face<'b>>

Create a face from the bytes of a given slice and an index specifying which font to read from an OpenType font collection.

pub fn from_table_func<'b, F>(func: F) -> Owned<Face<'b>>
where F: 'b + Send + Sync + FnMut(Tag) -> Option<Shared<Blob<'b>>>,

Create a new face from a closure that returns a raw Blob of table data.

pub fn face_data(&self) -> Shared<Blob<'a>>

pub fn table_with_tag(&self, tag: impl Into<Tag>) -> Option<Shared<Blob<'a>>>

Returns the slice of bytes for the table named tag or None if there is no table with tag.

pub fn index(&self) -> u32

pub fn set_upem(&mut self, upem: u32)

pub fn upem(&self) -> u32

pub fn set_glyph_count(&mut self, count: u32)

pub fn glyph_count(&self) -> u32

Returns the number of glyphs contained in the face.