harfbuzz_rs

Struct GlyphBuffer

pub struct GlyphBuffer(/* private fields */);

A GlyphBuffer contains the resulting output information of the shaping process.

An object of this type is obtained through the shape function.

Implementations

impl GlyphBuffer

pub fn len(&self) -> usize

Returns the length of the data of the buffer.

When called before shaping this is the number of unicode codepoints contained in the buffer. When called after shaping it returns the number of glyphs stored.

pub fn into_raw(self) -> *mut hb_buffer_t

Converts this buffer to a raw harfbuzz object pointer.

pub fn is_empty(&self) -> bool

Returns true if the buffer contains no elements.

pub fn get_glyph_positions(&self) -> &[GlyphPosition]

Get the glyph positions.

pub fn get_glyph_infos(&self) -> &[GlyphInfo]

Get the glyph infos.

pub fn reverse(&mut self)

Reverse the Buffer’s contents.

pub fn reverse_range(&mut self, start: usize, end: usize)

Reverse the Buffer’s contents in the range from start to end.

pub fn clear(self) -> UnicodeBuffer

Clears the contents of the glyph buffer and returns an empty UnicodeBuffer reusing the existing allocation.

pub fn serializer<'a>( &'a self, font: Option<&'a Font<'a>>, format: SerializeFormat, flags: SerializeFlags, ) -> BufferSerializer<'a>

Returns a serializer that allows the contents of the buffer to be converted into a human or machine readable representation.

Arguments

font: Optionally a font can be provided for access to glyph names and glyph extents. If None is passed an empty font is assumed.

format: The serialization format to use.

flags: Allows you to control which information will be contained in the serialized output.

Examples

Serialize the glyph buffer contents to a string using the textual format without any special flags.

use harfbuzz_rs::*;
use std::io::Read;
let face = Face::from_file(path, 0).expect("Error reading font file.");
let font = Font::new(face);

let buffer = UnicodeBuffer::new().add_str("ABC");

let buffer = shape(&font, buffer, &[]);

let mut string = String::new();
buffer
    .serializer(
        Some(&font),
        SerializeFormat::Text,
        SerializeFlags::default(),
    ).read_to_string(&mut string)
    .unwrap();

assert!(!string.is_empty()) // `string` contains the serialized values